+++ src/main/java/com/onsil/onsil/product/dto/ReviewDto.java
@@ -12,12 +12,12 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ReviewDto {
-    private Integer id;            // 리뷰 ID
-    private Integer productId;     // 상품 ID
-    private Integer memberId;      // 작성자 멤버 ID
-    private String memberName;     // 작성자 이름 (Member.userName)
-    private String content;        // 리뷰 내용
-    private int rating;            // 평점 (1~5)
-    private LocalDateTime regDate; // 작성일
-    private String image;          // 리뷰 이미지 (있으면)
+    private Integer id;
+    private Integer productId;
+    private Integer memberId;
+    private String memberName;
+    private String content;
+    private int rating;
+    private LocalDateTime regDate;
+    private String image;
 }
