+++ src/main/resources/templates/upload/products.html
@@ -5,7 +5,7 @@
     <title>Title</title>
 </head>
 <body>
-<form method="post" th:action="@{/product/save}" enctype="multipart/form-data">
+<form method="post" th:action="@{/upload/save}" enctype="multipart/form-data">
     <label for="flowerName">꽃 이름:</label>
     <input type="text" id="flowerName" name="flowerName" /><br/>
